requirements.txt
ffmpeg-python>=0.2.0
numpy>=1.20.0
pytest>=7.0.0
librosa>=0.10.0
soundfile>=0.12.1
mir_eval>=0.7
